requirements_fastapi.txt
Pillow==10.2.0
fastapi==0.101.1
numpy==1.24.3
uvicorn==0.23.2
scikit-learn==1.2.2
